requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Jinja2==3.1.6
python-dotenv==1.1.1
SQLAlchemy==2.0.43
Werkzeug==3.1.3
qrcode==8.2
pillow==11.3.0